+++ fractal_tasks_core/channels.py
@@ -18,8 +18,8 @@ from typing import Optional
 from typing import Union
 
 import zarr
-from pydantic import BaseModel
-from pydantic import validator
+from pydantic.v1 import BaseModel
+from pydantic.v1 import validator
 
 from fractal_tasks_core import __OME_NGFF_VERSION__
 
@@ -109,7 +109,9 @@ class ChannelInputModel(BaseModel):
 
     Attributes:
         wavelength_id: Unique ID for the channel wavelength, e.g. `A01_C01`.
-        label: Name of the channel.
+            Can only be specified if label is not set.
+        label: Name of the channel. Can only be specified if wavelength_id is
+            not set.
     """
 
     wavelength_id: Optional[str] = None