+++ fractal_tasks_core/dev/lib_signature_constraints.py
@@ -15,10 +15,10 @@ from inspect import signature
 from pathlib import Path
 from typing import Callable
 
-from pydantic.decorator import ALT_V_ARGS
-from pydantic.decorator import ALT_V_KWARGS
-from pydantic.decorator import V_DUPLICATE_KWARGS
-from pydantic.decorator import V_POSITIONAL_ONLY_NAME
+from pydantic.v1.decorator import ALT_V_ARGS
+from pydantic.v1.decorator import ALT_V_KWARGS
+from pydantic.v1.decorator import V_DUPLICATE_KWARGS
+from pydantic.v1.decorator import V_POSITIONAL_ONLY_NAME
 
 FORBIDDEN_PARAM_NAMES = (
     "args",
@@ -34,6 +34,7 @@ def _extract_function(
     module_relative_path: str,
     function_name: str,
     package_name: str = "fractal_tasks_core",
+    verbose: bool = False,
 ) -> Callable:
     """
     Extract function from a module with the same name.
@@ -42,6 +43,7 @@ def _extract_function(
         package_name: Example `fractal_tasks_core`.
         module_relative_path: Example `tasks/create_ome_zarr.py`.
         function_name: Example `create_ome_zarr`.
+        verbose:
     """
     if not module_relative_path.endswith(".py"):
         raise ValueError(f"{module_relative_path=} must end with '.py'")
@@ -49,8 +51,20 @@ def _extract_function(
         Path(module_relative_path).with_suffix("")
     )
     module_relative_path_dots = module_relative_path_no_py.replace("/", ".")
-    module = import_module(f"{package_name}.{module_relative_path_dots}")
-    task_function = getattr(module, function_name)
+    if verbose:
+        logging.info(
+            f"Now calling `import_module` for "
+            f"{package_name}.{module_relative_path_dots}"
+        )
+    imported_module = import_module(
+        f"{package_name}.{module_relative_path_dots}"
+    )
+    if verbose:
+        logging.info(
+            f"Now getting attribute {function_name} from "
+            f"imported module {imported_module}."
+        )
+    task_function = getattr(imported_module, function_name)
     return task_function