+++ fractal_tasks_core/tables/v1.py
@@ -11,9 +11,9 @@ from typing import Optional
 import anndata as ad
 import zarr.hierarchy
 from anndata.experimental import write_elem
-from pydantic import BaseModel
-from pydantic import validator
-from pydantic.error_wrappers import ValidationError
+from pydantic.v1 import BaseModel
+from pydantic.v1 import validator
+from pydantic.v1.error_wrappers import ValidationError
 
 from fractal_tasks_core.zarr_utils import OverwriteNotAllowedError