+++ fractal_tasks_core/tasks/copy_ome_zarr_hcs_plate.py
@@ -16,7 +16,7 @@ import logging
 from typing import Any
 
 import zarr
-from pydantic.decorator import validate_arguments
+from pydantic.v1.decorator import validate_arguments
 
 import fractal_tasks_core
 from fractal_tasks_core.ngff.specs import NgffPlateMeta