+++ fractal_tasks_core/tasks/init_group_by_well_for_multiplexing.py
@@ -14,7 +14,7 @@ Applies the multiplexing translation to all ROI tables
 """
 import logging
 
-from pydantic.decorator import validate_arguments
+from pydantic.v1.decorator import validate_arguments
 
 from fractal_tasks_core.tasks._registration_utils import (
     create_well_acquisition_dict,