+++ tests/dev/test_unit_lib_descriptions.py
@@ -10,19 +10,21 @@ from fractal_tasks_core.dev.lib_descriptions import (
 
 def test_get_function_args_descriptions():
     args_descriptions = _get_function_args_descriptions(
-        "fractal_tasks_core",
-        "dev/lib_signature_constraints.py",
-        "_extract_function",
+        package_name="fractal_tasks_core",
+        module_path="dev/lib_signature_constraints.py",
+        function_name="_extract_function",
     )
     debug(args_descriptions)
     assert args_descriptions.keys() == set(
-        ("package_name", "module_relative_path", "function_name")
+        ("package_name", "module_relative_path", "function_name", "verbose")
     )
 
 
 def test_get_class_attrs_descriptions():
     attrs_descriptions = _get_class_attrs_descriptions(
-        "fractal_tasks_core", "channels.py", "ChannelInputModel"
+        package_name="fractal_tasks_core",
+        module_relative_path="channels.py",
+        class_name="ChannelInputModel",
     )
     debug(attrs_descriptions)
     assert attrs_descriptions.keys() == set(("wavelength_id", "label"))